Fixed reCaptcha script loading
