+++ blocks/form/integrations/recaptcha.js
@@ -10,36 +10,6 @@ export default class GoogleReCaptcha {
     this.id = id;
   }
 
-  #loadScript(url) {
-    if (!this.loadPromise) {
-      this.loadPromise = new Promise((resolve, reject) => {
-        const head = document.head || document.querySelector('head');
-        const script = document.createElement('script');
-        script.src = url;
-        script.async = true;
-        script.onload = () => resolve(window.grecaptcha);
-        script.onerror = () => reject(new Error(`Failed to load script ${url}`));
-        head.append(script);
-      });
-    }
-  }
-
-  loadCaptcha(form) {
-    if (form && this.siteKey) {
-      const submit = form.querySelector('button[type="submit"]');
-      if (!submit) return;
-      const obs = new IntersectionObserver((entries) => {
-        entries.forEach((entry) => {
-          if (entry.isIntersecting) {
-            this.#loadScript(`https://www.google.com/recaptcha/api.js?render=${this.siteKey}`);
-            obs.disconnect();
-          }
-        });
-      });
-      obs.observe(submit);
-    }
-  }
-
   async getToken() {
     if (!this.siteKey) {
       return null;
